+++ msg/mavlink_log.msg
@@ -3,4 +3,4 @@ uint64 timestamp		# time since system start (microseconds)
 char[127] text
 uint8 severity # log level (same as in the linux kernel, starting with 0)
 
-uint8 ORB_QUEUE_LENGTH = 4
+uint8 ORB_QUEUE_LENGTH = 8
